[api] fixed QueryTags passing/failing at seemingly random
- api.SetTimestamps now preserves existing timestamps in archive if given an empty field

QueryTags would return a list of archive_id's sorted by the imported timestamp in descending order.
however, because the test was importing at roughly the same time, the import timestamp would be exactly the same value, causing it to fail randomly when randomTimestamps would round the timestamp to the nearest second.
thus at specific time intervals when the test gets ran, you could sometimes get a passing test, or a failing test whether or not you get lucky with timings of the test.

i have no idea why this bug seemed more reproducible on a linux machine (failing the test on the first initial run, then passing on all others), but not on windows. clearing test cache seems to have helped reproduce the bug only partially.